Batch commit for refactoring..
1.Basic imports are in basic_imports .py
2.Offtarget functions are in getofftargets.py
3.constants.py contains constants offcourse
4.parser_functions.py was created to separate out the argument parsing steps
5.bio_functions is basically responsible for biology related tasks..
6.similarly common_functions.py and fasta_functions.py are created..

I am going to further refactor the code in upcoming commits..
