+++ crispor.py
@@ -10,529 +10,26 @@ import Cookie, time, sys, cgi, re, random, platform, os
 import hashlib, base64, string, operator, urllib, sqlite3, time
 import traceback, json, pwd, pickle
 import common_functions
-import constants as cnst
-from datetime import datetime
+import parser_functions
+import get_offtargets
+
+from bio_functions import *
+from fasta_functions import *
+from common_functions import *
+from basic_imports import *
+from constants import *
 from collections import defaultdict, namedtuple
-from os.path import join, isfile, basename, dirname, isdir, abspath
-from StringIO import StringIO
+from datetime import datetime
 from itertools import product
+from os.path import join, isfile,  dirname, isdir, abspath
+from StringIO import StringIO
+ 
 
-# try to load external dependencies
-# we're going into great lengths to create a readable error message
-needModules = set(["tabix", "twobitreader", "pandas", "matplotlib", "scipy"])
-try:
-    import tabix # if not found, install with 'pip install pytabix'
-    needModules.remove("tabix") 
-except:
-    pass
-
-try:
-    import twobitreader # if not found, install with 'pip install twobitreader'
-    needModules.remove("twobitreader")
-except:
-    pass
-    
-try:
-    import pandas # required by doench2016 score. install with 'pip install pandas'
-    needModules.remove("pandas")
-    import scipy # required by doench2016 score. install with 'pip install pandas'
-    needModules.remove("scipy")
-    import matplotlib # required by doench2016 score. install with 'pip install matplotlib'
-    needModules.remove("matplotlib")
-    import numpy # required by doench2016 score. install with 'pip install numpy'
-    needModules.remove("numpy")
-except:
-    pass
-
-if len(needModules)!=0:
-    print("Content-type: text/html\n")
-    print("Python interpreter path: %s<p>" % sys.executable)
-    print("These python modules were not found: %s<p>" % ",".join(needModules))
-    print("To install all requirements in one line, run: pip install pytabix pandas twobitreader scipy matplotlib numpy<p>")
-    sys.exit(0)
-
-# our own eff scoring library
-import crisporEffScores
-
-# don't report print as an error
-# pylint: disable=E1601
-
-# optional module for Excel export as native .xls files
-# install with 'apt-get install python-xlwt' or 'pip install xlwt'
-xlwtLoaded = True
-try:
-    import xlwt
-except:
-    sys.stderr.write("crispor.py - warning - the python xlwt module is not available\n")
-    xlwtLoaded = False
-
-# optional module for mysql support
-try:
-    import MySQLdb
-    mysqldbLoaded = True
-except:
-    mysqldbLoaded = False
-
-# version of crispor
-versionStr = "4.0"
-
-# contact email
-contactEmail='crispor@tefor.net'
-
-# write debug output to stdout
-DEBUG = False
-#DEBUG = True
-
-# use bowtie for off-target search?
-useBowtie = False
-
-# calculate the efficienc scores?
-doEffScoring = True
-
-# system-wide temporary directory
-TEMPDIR = os.environ.get("TMPDIR", "/tmp")
-#TEMPDIR = "/dev/shm/"
-
-# prefix in html statements before the directories "image/", "style/" and "js/" 
-HTMLPREFIX =  ""
-# alternative directory on local disk where image/, style/ and js/ are located
-HTMLDIR = "/usr/local/apache/htdocs/crispor/"
-
-# directory of crispor.py
-baseDir = cnst.baseDir
-# filename of this script, usually crispor.py
-myName = basename(__file__)
-
-# the segments.bed files use abbreviated genomic region names
-segTypeConv = {"ex":"exon", "in":"intron", "ig":"intergenic"}
-
-# directory for processed batches of offtargets ("cache" of bwa results)
-batchDir = join(baseDir,"temp")
-
-# use mysql or sqlite for the jobs?
-jobQueueUseMysql = False
-
-# the file where the sqlite job queue is stored
-JOBQUEUEDB = join("/tmp/crisporJobs.db")
-
-# alternatively: connection info for mysql
-jobQueueMysqlConn = {"socket":None, "host":None, "user": None, "password" : None}
-
-# directory for platform-independent scripts (e.g. Heng Li's perl SAM parser)
-scriptDir = join(baseDir, "bin")
-
-# directory for helper binaries (e.g. BWA)
-binDir = abspath(join(baseDir, "bin", platform.system()))
-
-
-
-DEFAULTORG = 'hg19'
-DEFAULTSEQ = 'cttcctttgtccccaatctgggcgcgcgccggcgccccctggcggcctaaggactcggcgcgccggaagtggccagggcgggggcgacctcggctcacagcgcgcccggctattctcgcagctcaccatgGATGATGATATCGCCGCGCTCGTCGTCGACAACGGCTCCGGCATGTGCAAGGCCGGCTTCGCGGGCGACGATGCCCCCCGGGCCGTCTTCCCCTCCATCGTGGGGCGCC'
-
-# used if hg19 is not available
-ALTORG = 'sacCer3'
-ALTSEQ = 'ATTCTACTTTTCAACAATAATACATAAACatattggcttgtggtagCAACACTATCATGGTATCACTAACGTAAAAGTTCCTCAATATTGCAATTTGCTTGAACGGATGCTATTTCAGAATATTTCGTACTTACACAGGCCATACATTAGAATAATATGTCACATCACTGTCGTAACACTCT'
-
-
-
-DEFAULTPAM = 'NGG'
-
-
-# maximum size of an input sequence 
-MAXSEQLEN = 1000
-# maximum input size when specifying "no genome"
-MAXSEQLEN_NOGENOME = 25000
-
-# BWA: allow up to X mismatches
-maxMMs=4
-
-# maximum number of occurences in the genome to get flagged as repeats. 
-# This is used in bwa samse, when converting the same file
-# and for warnings in the table output.
-MAXOCC = 60000
-
-# the BWA queue size is 2M by default. We derive the queue size from MAXOCC
-MFAC = 2000000/MAXOCC
-
-# the length of the guide sequence, set by setupPamInfo
-GUIDELEN=None
-
-# the name of the currently processed batch, assigned only once 
-# in readBatchParams and only for json-type batches
-batchName = ""
-
-# are we doing a Cpf1 run?
-# this variable changes almost all processing and
-# has to be set on program start, as soon as we know 
-# the PAM we're running on
-cpf1Mode=None
-
-
-# Highly-sensitive mode (not for CLI mode): 
-# MAXOCC is increased in processSubmission() and in the html UI if only one
-# guide seq is run
-# Also, the number of allowed mismatches is increased to 5 instead of 4
-#HIGH_MAXOCC=600000
-#HIGH_maxMMs=5
-
-# minimum off-target score of standard off-targets (those that end with NGG)
-# This should probably be based on the CFD score these days
-# But for now, I'll let the user do the filtering
-MINSCORE = 0.0
-
-
-
-# for some PAMs, we allow other alternative motifs when searching for offtargets
-# MIT and eCrisp do that, they use the motif NGG + NAG, we add one more, based on the
-# on the guideSeq results in Tsai et al, Nat Biot 2014
-# The NGA -> NGG rule was described by Kleinstiver...Young 2015 "Improved Cas9 Specificity..."
-# NNGTRRT rule for S. aureus is in the new protocol "SaCas9 User manual"
-# ! the length of the alternate PAM has to be the same as the original PAM!
-offtargetPams = {
-"NGG" : ["NAG","NGA"],
-"NGA" : ["NGG"],
-"NNGRRT" : ["NNGRRN"]
-}
-
-# global flag to indicate if we're run from command line or as a CGI
-commandLineMode = False
-
-# names/order of efficiency scores to show in UI
-scoreNames = ["fusi", "crisprScan"]
-allScoreNames = ["fusi", "chariRank", "ssc", "doench", "wang", "crisprScan", "oof", "housden", "proxGc"]
-
-# how many digits shall we show for each score? default is 0
-scoreDigits = {
-    "ssc" : 1,
-}
-
-# List of AddGene plasmids, their long and short names:
-addGenePlasmids = [
-("43860", ("MLM3636 (Joung lab)", "MLM3636")),
-("49330", ("pAc-sgRNA-Cas9 (Liu lab)", "pAcsgRnaCas9")),
-("42230", ("pX330-U6-Chimeric_BB-CBh-hSpCas9 (Zhang lab) + derivatives", "pX330")),
-("52961", ("lentiCRISPR v2 (Zhang lab)", "lentiCrispr")),
-]
-
-addGenePlasmidsAureus = [
-("61591", ("pX601-AAV-CMV::NLS-SaCas9-NLS-3xHA-bGHpA;U6::BsaI-sgRNA (Zhang lab)", "pX601")),
-("61592", ("pX600-AAV-CMV::NLS-SaCas9-NLS-3xHA-bGHpA (Zhang lab)", "pX600")),
-("61593", ("pX602-AAV-TBG::NLS-SaCas9-NLS-HA-OLLAS-bGHpA;U6::BsaI-sgRNA (Zhang lab)", "pX602")),
-("65779", ("VVT1 (Joung lab)", "VVT1"))
-]
-
-# list of AddGene primer 5' and 3' extensions, one for each AddGene plasmid
-# format: prefixFw, prefixRw, restriction enzyme, link to protocol
-addGenePlasmidInfo = {
-"43860" : ("ACACC", "AAAAC", "BsmBI", "https://www.addgene.org/static/data/plasmids/43/43860/43860-attachment_T35tt6ebKxov.pdf"),
-"49330" : ("TTC", "AAC", "Bsp QI", "http://bio.biologists.org/content/3/1/42#sec-9"),
-"42230" : ("CACC", "AAAC", "Bbs1", "https://www.addgene.org/static/data/plasmids/52/52961/52961-attachment_B3xTwla0bkYD.pdf"),
-"52961" : ("CACC", "AAAC", "Bbs1", "https://www.addgene.org/static/data/plasmids/52/52961/52961-attachment_B3xTwla0bkYD.pdf"),
-"61591" : ("CACC", "AAAC", "BsaI", "https://www.addgene.org/static/data/plasmids/61/61591/61591-attachment_it03kn5x5O6E.pdf"),
-"61592" : ("CACC", "AAAC", "BsaI", "https://www.addgene.org/static/data/plasmids/61/61592/61592-attachment_iAbvIKnbqNRO.pdf"),
-"61593" : ("CACC", "AAAC", "BsaI", "https://www.addgene.org/static/data/plasmids/61/61592/61592-attachment_iAbvIKnbqNRO.pdf"),
-"65779": ("CACC", "AAAC", "BsmBI", "https://www.addgene.org/static/data/plasmids/65/65779/65779-attachment_G8oNyvV6pA78.pdf")
-}
-
-# Restriction enzyme supplier codes
-rebaseSuppliers = {
-"B":"Life Technologies",
-"C":"Minotech",
-"E":"Agilent",
-"I":"SibEnzyme",
-"J":"Nippon Gene",
-"K":"Takara",
-"M":"Roche",
-"N":"NEB",
-"O":"Toyobo",
-"Q":"Molecular Biology Resources",
-"R":"Promega",
-"S":"Sigma",
-"V":"Vivantis",
-"X":"EURx",
-"Y":"SinaClon BioScience"
-}
-
-# labels and descriptions of eff. scores
-scoreDescs = {
-    "doench" : ("Doench '14", "Range: 0-100. Linear regression model trained on 880 guides transfected into human MOLM13/NB4/TF1 cells (three genes) and mouse cells (six genes). Delivery: lentivirus. The Fusi score can be considered an updated version this score, as their training data overlaps a lot. See <a target='_blank' href='http://www.nature.com/nbt/journal/v32/n12/full/nbt.3026.html'>Doench et al.</a>"),
-    "ssc" : ("Xu", "Range ~ -2 - +2. Aka 'SSC score'. Linear regression model trained on data from &gt;1000 genes in human KBM7/HL60 cells (Wang et al) and mouse (Koike-Yusa et al.). Delivery: lentivirus. Ranges mostly -2 to +2. See <a target='_blank' href='http://genome.cshlp.org/content/early/2015/06/10/gr.191452.115'>Xu et al.</a>"),
-    "crisprScan" : ["Moreno-Mateos", "Also called 'CrisprScan'. Range: mostly 0-100. Linear regression model, trained on data from 1000 guides on &gt;100 genes, from zebrafish 1-cell stage embryos injected with mRNA. See <a target=_blank href='http://www.nature.com/nmeth/journal/v12/n10/full/nmeth.3543.html'>Moreno-Mateos et al.</a>. Recommended for guides transcribed <i>in-vitro</i> (T7 promoter). Click to sort by this score."],
-    "wang" : ("Wang", "Range: 0-100. SVM model trained on human cell culture data on guides from &gt;1000 genes. The Xu score can be considered an updated version of this score, as the training data overlaps a lot. Delivery: lentivirus. See <a target='_blank' href='http://http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3972032/'>Wang et al.</a>"),
-    "chariRank" : ("Chari", "Range: 0-100. Support Vector Machine, converted to rank-percent, trained on data from 1235 guides targeting sequences that were also transfected with a lentivirus into human 293T cells. See <a target='_blank' href='http://www.nature.com/nmeth/journal/v12/n9/abs/nmeth.3473.html'>Chari et al.</a>"),
-    "fusi" : ("Doench '16", "Previously called the 'Fusi' score. Range: 0-100. Boosted Regression Tree model, trained on data produced by Doench et al (881 guides, MOLM13/NB4/TF1 cells + unpublished additional data). Delivery: lentivirus. See <a target='_blank' href='http://biorxiv.org/content/early/2015/06/26/021568'>Fusi et al. 2015</a> and <a target='_blank' href='http://www.nature.com/nbt/journal/v34/n2/full/nbt.3437.html'>Doench et al. 2016</a>. Recommended for guides expressed in cells (U6 promoter). Click to sort the table by this score."),
-    "housden" : ("Housden", "Range: ~ 1-10. Weight matrix model trained on data from Drosophila mRNA injections. See <a target='_blank' href='http://stke.sciencemag.org/content/8/393/rs9.long'>Housden et al.</a>"),
-    "oof" : ("Out-of-Frame", "Range: 0-100. Predicts the percentage of clones that will carry out-of-frame deletions, based on the micro-homology in the sequence flanking the target site. See <a target='_blank' href='http://www.nature.com/nmeth/journal/v11/n7/full/nmeth.3015.html'>Bae et al.</a>. Click the score to show the most likely deletions for this guide.")
-}
-
-# the headers for the guide and offtarget output files
-guideHeaders = ["guideId", "guideSeq", "mitSpecScore", "cfdSpecScore", "offtargetCount", "guideGenomeMatchGeneLocus"]
-offtargetHeaders = ["guideId", "guideSeq", "offtargetSeq", "mismatchPos", "mismatchCount", "mitOfftargetScore", "cfdOfftargetScore", "chrom", "start", "end", "strand", "locusDesc"]
-
-# a file crispor.conf in the directory of the script allows to override any global variable
-myDir = dirname(__file__)
-confPath =join(myDir, "crispor.conf")
-if isfile(confPath):
-    exec(open(confPath))
-
-cgiParams = None
-
-# ====== END GLOBALS ============
-
-def setupPamInfo(pam):
-    " modify a few globals based on the current pam "
-    global GUIDELEN
-    global cpf1Mode
-    global addGenePlasmids
-    if pam=="TTTN":
-        GUIDELEN = 23
-        cpf1Mode = True
-    elif pam=="NNGRRT":
-        addGenePlasmids = addGenePlasmidsAureus
-        GUIDELEN = 21
-        cpf1Mode = False
-    else:
-        GUIDELEN = 20
-        cpf1Mode = False
 
 
 
 # ====== FUNCTIONS =====
 contentLineDone = False
-
-def errAbort(msg):
-    " print err msg and exit "
-    if commandLineMode:
-        raise Exception(msg)
-    sys.exit(0)  # cgi must not exit with 1
-
-# allow only dashes, digits, characters, underscores and colons in the CGI parameters
-# and +
-notOkChars = re.compile(r'[^a-zA-Z0-9-_:+.]')
-
-def checkVal(key, str):
-    """ remove special characters from input string, to protect against injection attacks """
-    if len(str) > 10000:
-	errAbort("input parameter %s is too long" % key)
-    if notOkChars.search(str):
-	errAbort("input parameter %s contains an invalid character" % key)
-    return str
-
-def cgiGetParams():
-    " get CGI parameters and return as dict "
-    form = cgi.FieldStorage()
-    global cgiParams
-    cgiParams = {}
-
-    # parameters are:
-    #"pamId", "batchId", "pam", "seq", "org", "download", "sortBy", "format", "ajax
-    for key in form.keys():
-        val = form.getfirst(key)
-	if val!=None:
-            # "seq" is cleaned by cleanSeq later
-            val = urllib.unquote(val)
-            if key not in ["seq", "name"]:
-                checkVal(key, val)
-            cgiParams[key] = val
-
-    if "pam" in cgiParams:
-        if len(set(cgiParams["pam"])-set("ACTGNMKRY"))!=0:
-            errAbort("Illegal character in PAM-sequence. Only ACTGMKRY and N allowed.")
-    return cgiParams
-
-transTab = string.maketrans("-=/+_", "abcde")
-
-def makeTempBase(seq, org, pam, batchName):
-    "create the base name of temp files using a hash function and some prettyfication "
-    hasher = hashlib.sha1(seq+org+pam+batchName)
-    batchId = base64.urlsafe_b64encode(hasher.digest()[0:20]).translate(transTab)[:20]
-    return batchId
-
-def makeTempFile(prefix, suffix):
-    " return a temporary file that is deleted upon exit, unless DEBUG is set "
-    if DEBUG:
-        fname = join("/tmp", prefix+suffix)
-        fh = open(fname, "w")
-    else:
-        fh = tempfile.NamedTemporaryFile(prefix="primer3In", suffix=".txt")
-    return fh
-
-
-
-def debug(msg):
-    if commandLineMode:
-        logging.debug(msg)
-    elif DEBUG:
-        print msg
-        print "<br>"
-
-def matchNuc(pat, nuc):
-    " returns true if pat (single char) matches nuc (single char) "
-    if pat in ["A", "C", "T", "G"] and pat==nuc:
-        return True
-    elif pat=="S" and nuc in ["G", "C"]:
-        return True
-    elif pat=="M" and nuc in ["A", "C"]:
-        return True
-    elif pat=="D" and nuc in ["AGT"]:
-        return True
-    elif pat=="W" and nuc in ["A", "T"]:
-        return True
-    elif pat=="K" and nuc in ["T", "G"]:
-        return True
-    elif pat=="R" and nuc in ["A", "G"]:
-        return True
-    elif pat=="Y" and nuc in ["C", "T"]:
-        return True
-    else:
-        return False
-
-def gcContent(seq):
-    " return GC content as a float "
-    c = 0
-    for x in seq:
-        if x in ["G","C"]:
-            c+= 1
-    return (float(c)/len(seq))
-            
-def findPat(seq, pat):
-    """ yield positions where pat matches seq, stupid brute force search 
-    """
-    seq = seq.upper()
-    pat = pat.upper()
-    for i in range(0, len(seq)-len(pat)+1):
-        #print "new pos", i, seq[i:i+len(pat)],"<br>"
-        found = True
-        for x in range(0, len(pat)):
-            #print "new step", x, "<br>"
-            if pat[x]=="N":
-                #print "N","<br>"
-                continue
-            seqPos = i+x
-            if seqPos == len(seq):
-                found = False
-                break
-            if not matchNuc(pat[x], seq[seqPos]):
-            #if not patMatch(seq[seqPos], pat[x]):
-                #print i, x, pat[x], seq[seqPos], "no match<br>"
-                found = False
-                break
-            #print "match", i, x, found, "<br>"
-        if found:
-            #print "yielding", i, "<br>"
-            yield i
-
-def rndSeq(seqLen):
-    " return random seq "
-    seq = []
-    alf = "ACTG"
-    for i in range(0, seqLen):
-        seq.append(alf[random.randint(0,3)])
-    return "".join(seq)
-
-def cleanSeq(seq, db):
-    """ remove fasta header, check seq for illegal chars and return (filtered
-    seq, user message) special value "random" returns a random sequence.
-    """
-    #print repr(seq)
-    if seq.startswith("random"):
-        seq = rndSeq(800)
-    lines = seq.strip().splitlines()
-    #print "<br>"
-    #print "before fasta cleaning", "|".join(lines)
-    if len(lines)>0 and lines[0].startswith(">"):
-        line1 = lines.pop(0)
-    #print "<br>"
-    #print "after fasta cleaning", "|".join(lines)
-    #print "<br>"
-
-    newSeq = []
-    nCount = 0
-    for l in lines:
-        if len(l)==0:
-            continue
-        for c in l:
-            if c not in "actgACTGNn":
-                nCount +=1
-            else:
-                newSeq.append(c)
-    seq = "".join(newSeq)
-
-    msgs = []
-    if len(seq)>MAXSEQLEN and db!="noGenome":
-        msgs.append("<strong>Sorry, this tool cannot handle sequences longer than 1kbp</strong><br>Below you find the results for the first %d bp of your input sequence.<br>" % MAXSEQLEN)
-        seq = seq[:MAXSEQLEN]
-    if len(seq)>MAXSEQLEN_NOGENOME and db=="noGenome":
-        msgs.append("<strong>Sorry, this tool cannot handle sequences longer than %d bp when specifying 'No Genome'.</strong><br>Below you find the results for the first %d bp of your input sequence.<br>" % (MAXSEQLEN_NOGENOME, MAXSEQLEN_NOGENOME))
-        seq = seq[:MAXSEQLEN_NOGENOME]
-
-    if nCount!=0:
-        msgs.append("Sequence contained %d non-ACTGN letters. They were removed." % nCount)
-
-    return seq, "<br>".join(msgs)
-
-def revComp(seq):
-    " rev-comp a dna sequence with UIPAC characters "
-    revTbl = {'A' : 'T', 'C' : 'G', 'G' : 'C', 'T' : 'A', 'N' : 'N' , 'M' : 'K', 'K' : 'M', 
-        "R" : "Y" , "Y":"R" , "g":"c", "a":"t", "c":"g","t":"a", "n":"n"}
-    newSeq = []
-    for c in reversed(seq):
-        newSeq.append(revTbl[c])
-    return "".join(newSeq)
-
-def findPams (seq, pam, strand, startDict, endSet):
-    """ return two values: dict with pos -> strand of PAM and set of end positions of PAMs
-    Makes sure to return only values with at least GUIDELEN bp left (if strand "+") or to the
-    right of the match (if strand "-")
-    If the PAM is TTTN, then this is inversed: pos-strand matches must have at least GUIDELEN
-    basepairs to the right, neg-strand matches must have at least GUIDELEN bp on their left
-    >>> findPams("GGGGGGGGGGGGGGGGGGGGGGG", "NGG", "+", {}, set(), False)
-    ({20: '+'}, set([23]))
-    >>> findPams("CCAGCCCCCCCCCCCCCCCCCCC", "CCA", "-", {}, set(), False)
-    ({0: '-'}, set([3]))
-    >>> findPams("TTTNCCCCCCCCCCCCCCCCCTTTN", "TTTN", "+", {}, set(), True)
-    ({0: '+'}, set([3]))
-    >>> findPams("CCCCCCCCCCCCCCCCCCCCCAAAA", "NAA", "-", {}, set(), True
-    ({}, set([]))
-    >>> findPams("AAACCCCCCCCCCCCCCCCCCCCC", "NAA", "-", {}, set(), True)
-    ({}, set([]))
-    >>> findPams("CCCCCCCCCCCCCCCCCCCCCCCCCAA", "NAA", "-", {}, set(), True)
-    ({}, set([]))
-
-    """
-    assert(cpf1Mode is not None)
-
-    if cpf1Mode:
-        maxPosPlus  = len(seq)-(GUIDELEN+len(pam))
-        minPosMinus = GUIDELEN
-    else:
-        # -------------------
-        #          OKOKOKOKOK
-        minPosPlus  = GUIDELEN
-        # -------------------
-        # OKOKOKOKOK
-        maxPosMinus = len(seq)-(GUIDELEN+len(pam))
-
-    #print "new search", seq, pam, "<br>"
-    for start in findPat(seq, pam):
-        if cpf1Mode:
-            # need enough flanking seq on one side
-            #print "found", start,"<br>"
-            if strand == "+" and start > maxPosPlus:
-                continue
-            if strand == "-" and start < minPosMinus:
-                continue
-        else:
-            if strand=="+" and start < minPosPlus:
-                continue
-            if strand=="-" and start > maxPosMinus:
-                continue
-
-        #print "match", strand, start, end, "<br>"
-        startDict[start] = strand
-        end = start+len(pam)
-        endSet.add(end)
-    return startDict, endSet
 
 def rulerString(maxLen):
     " return line with positions every 10 chars "
@@ -559,12 +56,14 @@ def iterOneDelSeqs(seq):
             yield i, delSeq
         doneSeqs.add(delSeq)
 
-def flankSeqIter(seq, startDict, pamLen, doFilterNs):
+def flankSeqIter(seq, startDict, pam, doFilterNs):
     """ given a seq and dictionary of pos -> strand and the length of the pamSite
     yield tuples of (name, pamStart, guideStart, strand, flankSeq, pamSeq)
 
     if doFilterNs is set, will not return any sequences that contain an N character
     """
+    pamLen = len(pam)
+    GUIDELEN,cpf1Mode,addGenePlasmids = common_functions.setupPamInfo(pam)
     startList = sorted(startDict.keys())
     for pamStart in startList:
         strand = startDict[pamStart]
@@ -904,7 +403,7 @@ def calcCfdScore(guideSeq, otSeq):
 
 # --- END OF SCORING ROUTINES 
 
-def parseChromSizes(genome):
+def parseChromSizes(genome,genomesDir):
     " return chrom sizes as dict chrom -> size "
     genomeDir = genomesDir # make local
     sizeFname = "%(genomeDir)s/%(genome)s/%(genome)s.sizes" % locals()
@@ -915,13 +414,13 @@ def parseChromSizes(genome):
         ret[chrom] = int(size)
     return ret
 
-def extendAndGetSeq(db, chrom, start, end, strand, flank=100):
+def extendAndGetSeq(db,genomesDir, chrom, start, end, strand, flank=100):
     """ extend (start, end) by flank and get sequence for it using twoBitTwoFa.
     Return None if not possible to extend.
     #>>> extendAndGetSeq("hg19", "chr21", 10000000, 10000005, "+", flank=3)
     #'AAGGAATGTAG'
     """
-    chromSizes = parseChromSizes(db)
+    chromSizes = parseChromSizes(db,genomesDir)
     maxEnd = chromSizes[chrom]+1
 
     start -= flank
@@ -993,7 +492,7 @@ def getExtSeq(seq, start, end, strand, extUpstream, extDownstream, extSeq=None, 
 
     return subSeq
 
-def pamStartToGuideRange(startPos, strand, pamLen):
+def pamStartToGuideRange(startPos, strand, pamLen,cpf1Mode,GUIDELEN):
     """ given a PAM start position and its strand, return the (start,end) of the guide.
     Coords can be negative or exceed the length of the input sequence.
     """
@@ -1122,7 +621,7 @@ def mergeGuideInfo(seq, startDict, pamPat, otMatches, inputPos, effScores, sortB
     guideScores = {}
     hasNotFound = False
 
-    pamSeqs = list(flankSeqIter(seq, startDict, len(pamPat), True))
+    pamSeqs = list(flankSeqIter(seq, startDict, pamPat, True))
 
     for pamId, pamStart, guideStart, strand, guideSeq, pamSeq in pamSeqs:
         # matches in genome
@@ -1292,26 +791,15 @@ def writePamFlank(seq, startDict, pam, faFname):
     " write pam flanking sequences to fasta file, optionally with versions where each nucl is removed "
     #print "writing pams to %s<br>" % faFname
     faFh = open(faFname, "w")
-    for pamId, pamStart, guideStart, strand, flankSeq, pamSeq in flankSeqIter(seq, startDict, len(pam), True):
+    for pamId, pamStart, guideStart, strand, flankSeq, pamSeq in flankSeqIter(seq, startDict, pam, True):
         faFh.write(">%s\n%s\n" % (pamId, flankSeq))
     faFh.close()
 
-def runCmd(cmd, ignoreExitCode=False):
-    " run shell command, check ret code, replaces BIN and SCRIPTS special variables "
-    cmd = cmd.replace("$BIN", binDir)
-    cmd = cmd.replace("$SCRIPT", scriptDir)
-    cmd = "set -o pipefail; " + cmd
-    debug("Running %s" % cmd)
-    ret = subprocess.call(cmd, shell=True, executable="/bin/bash")
-    if ret!=0 and not ignoreExitCode:
-        if commandLineMode:
-            logging.error("Error: could not run command %s." % cmd)
-            sys.exit(1)
-        else:
-            print "Server error: could not run command %s, error %d.<p>" % (cmd, ret)
-            print "please send us an email, we will fix this error as quickly as possible. %s " % contactEmail
-            sys.exit(0)
 
+def debug(msg):
+    logging.debug(msg)
+    if DEBUG:
+        print msg
 
 def parseOfftargets(bedFname):
     """ parse a bed file with annotataed off target matches from overlapSelect,
@@ -1405,19 +893,20 @@ def calcGuideEffScores(seq, extSeq, pam):
 
     extSeq can be None, if we were unable to extend the sequence
     """
+    GUIDELEN,cpf1Mode,addGenePlasmids = common_functions.setupPamInfo(pam)
     seq = seq.upper()
     if extSeq:
         extSeq = extSeq.upper()
 
     startDict, endSet = findAllPams(seq, pam)
-    pamInfo = list(flankSeqIter(seq, startDict, len(pam), False))
+    pamInfo = list(flankSeqIter(seq, startDict, pam, False))
 
     guideIds = []
     guides = []
     longSeqs = []
     for pamId, startPos, guideStart, strand, guideSeq, pamSeq in pamInfo:
         guides.append(guideSeq+pamSeq)
-        gStart, gEnd = pamStartToGuideRange(startPos, strand, len(pam))
+        gStart, gEnd = pamStartToGuideRange(startPos, strand, len(pam),cpf1Mode,GUIDELEN)
         longSeq = getExtSeq(seq, gStart, gEnd, strand, 50-GUIDELEN, 50, extSeq)
         if longSeq!=None:
             longSeqs.append(longSeq)
@@ -1500,66 +989,7 @@ def readEffScores(batchId):
         seqToScores[row.guideId] = scoreDict
     return seqToScores
 
-def findOfftargetsBwa(queue, batchId, batchBase, faFname, genome, pam, bedFname):
-    " align faFname to genome and create matchedBedFname "
-    matchesBedFname = batchBase+".matches.bed"
-    saFname = batchBase+".sa"
-    pamLen = len(pam)
-    genomeDir = genomesDir # make var local, see below
 
-    open(matchesBedFname, "w") # truncate to 0 size
-
-    # increase MAXOCC if there is only a single query, but only in CGI mode
-    #if len(parseFasta(open(faFname)))==1 and not commandLineMode:
-        #global MAXOCC
-        #global maxMMs
-        #MAXOCC=max(HIGH_MAXOCC, MAXOCC)
-        #maxMMs=HIGH_maxMMs
-
-    maxDiff = maxMMs
-    queue.startStep(batchId, "bwa", "Alignment of potential guides, mismatches <= %d" % maxDiff)
-    convertMsg = "Converting alignments"
-    seqLen = GUIDELEN
-
-    bwaM = MFAC*MAXOCC # -m is queue size in bwa
-    cmd = "$BIN/bwa aln -o 0 -m %(bwaM)s -n %(maxDiff)d -k %(maxDiff)d -N -l %(seqLen)d %(genomeDir)s/%(genome)s/%(genome)s.fa %(faFname)s > %(saFname)s" % locals()
-    runCmd(cmd)
-
-    queue.startStep(batchId, "saiToBed", convertMsg)
-    maxOcc = MAXOCC # create local var from global
-    # EXTRACTION OF POSITIONS + CONVERSION + SORT/CLIP
-    # the sorting should improve the twoBitToFa runtime
-    cmd = "$BIN/bwa samse -n %(maxOcc)d %(genomeDir)s/%(genome)s/%(genome)s.fa %(saFname)s %(faFname)s | $SCRIPT/xa2multi.pl | $SCRIPT/samToBed %(pam)s | sort -k1,1 -k2,2n | $BIN/bedClip stdin %(genomeDir)s/%(genome)s/%(genome)s.sizes stdout >> %(matchesBedFname)s " % locals()
-    runCmd(cmd)
-
-    # arguments: guideSeq, mainPat, altPats, altScore, passX1Score
-    filtMatchesBedFname = batchBase+".filtMatches.bed"
-    queue.startStep(batchId, "filter", "Removing matches without a PAM motif")
-    altPats = ",".join(offtargetPams.get(pam, ["na"]))
-    bedFnameTmp = bedFname+".tmp"
-    altPamMinScore = str(ALTPAMMINSCORE)
-    # EXTRACTION OF SEQUENCES + ANNOTATION
-    # twoBitToFa was 15x slower than python's twobitreader, after markd's fix it should be OK
-    cmd = "$BIN/twoBitToFa %(genomeDir)s/%(genome)s/%(genome)s.2bit stdout -bed=%(matchesBedFname)s | $SCRIPT/filterFaToBed %(faFname)s %(pam)s %(altPats)s %(altPamMinScore)s %(maxOcc)d > %(filtMatchesBedFname)s" % locals()
-    #cmd = "$SCRIPT/twoBitToFaPython %(genomeDir)s/%(genome)s/%(genome)s.2bit %(matchesBedFname)s | $SCRIPT/filterFaToBed %(faFname)s %(pam)s %(altPats)s %(altPamMinScore)s %(maxOcc)d > %(filtMatchesBedFname)s" % locals()
-    runCmd(cmd)
-
-    segFname = "%(genomeDir)s/%(genome)s/%(genome)s.segments.bed" % locals()
-
-    # if we have gene model segments, annotate them, otherwise just use the chrom position
-    if isfile(segFname):
-        queue.startStep(batchId, "genes", "Annotating matches with genes")
-        cmd = "cat %(filtMatchesBedFname)s | $BIN/overlapSelect %(segFname)s stdin stdout -mergeOutput -selectFmt=bed -inFmt=bed | cut -f1,2,3,4,8 2> %(batchBase)s.log > %(bedFnameTmp)s " % locals()
-        runCmd(cmd)
-    else:
-        queue.startStep(batchId, "chromPos", "Annotating matches with chromosome position")
-        annotateBedWithPos(filtMatchesBedFname, bedFnameTmp)
-
-    # make sure the final bed file is never in a half-written state, 
-    # as it is our signal that the job is complete
-    shutil.move(bedFnameTmp, bedFname)
-    queue.startStep(batchId, "done", "Job completed")
-    return bedFname
 
 def makeVariants(seq):
     " generate all possible variants of sequence at 1bp-distance"
@@ -1690,129 +1120,11 @@ def getEditDist(str1, str2):
             editDist +=1
     return editDist
 
-def findOfftargetsBowtie(queue, batchId, batchBase, faFname, genome, pamPat, bedFname):
-    " align guides with pam in faFname to genome and write off-targets to bedFname "
-    tmpDir = batchBase+".bowtie.tmp"
-    os.mkdir(tmpDir)
-
-    # make sure this directory gets removed, no matter what
-    global tmpDirsDelExit
-    tmpDirsDelExit.append(tmpDir)
-    if not DEBUG:
-        atexit.register(delTmpDirs)
-
-    # write out the sequences for bowtie
-    queue.startStep(batchId, "seqPrep", "preparing sequences")
-    bwFaFname = abspath(join(tmpDir, "bowtieIn.fa"))
-    guideSeqs, qSeqs, allPamSeqs = writeBowtieSequences(faFname, bwFaFname, pamPat)
-
-    genomePath =  abspath(join(genomesDir, genome, genome))
-    oldCwd = os.getcwd()
-
-    # run bowtie
-    queue.startStep(batchId, "bowtie", "aligning with bowtie")
-    os.chdir(tmpDir) # bowtie writes to hardcoded output filenames with --refout
-    # -v 3 = up to three mismatches
-    # -y   = try hard
-    # -t   = print time it took
-    # -k   = output up to X alignments
-    # -m   = do not output any hit if a read has more than X hits
-    # --max = write all reads that exceed -m to this file
-    # --refout = output in bowtie format, not SAM
-    # --maxbts=2000 maximum number of backtracks
-    # -p 4 = use four threads
-    # --mm = use mmap
-    maxOcc = MAXOCC # meaning in BWA: includes any PAM, in bowtie we have the PAM in the input sequence
-    cmd = "$BIN/bowtie %(genomePath)s -f %(bwFaFname)s  -v 3 -y -t -k %(maxOcc)d -m %(maxOcc)d dummy --max tooManyHits.txt --mm --refout --maxbts=2000 -p 4" % locals()
-    runCmd(cmd)
-    os.chdir(oldCwd)
-
-    queue.startStep(batchId, "parse", "parsing alignments")
-    pamLen = len(pamPat)
-    hits = parseRefout(tmpDir, guideSeqs, pamLen)
-
-    queue.startStep(batchId, "scoreOts", "scoring off-targets")
-    # make the list of alternative PAM sequences
-    altPats = offtargetPams.get(pamPat, [])
-    altPamSeqs = []
-    for altPat in altPats:
-        altPamSeqs.extend(expandIupac(altPat))
-
-    # iterate over bowtie hits and write to a BED file with scores
-    # if the hit looks OK (right PAM + score is high enough)
-    tempBedPath = join(tmpDir, "bowtieHits.bed")
-    tempFh = open(tempBedPath, "w")
-
-    offTargets = {}
-    for guideIdWithMod, chrom, start, end, strand, _, tSeq in hits:
-        guideId = guideIdWithMod.split(".")[0]
-        guideSeq = guideSeqs[guideId]
-        genomePamSeq = tSeq[-pamLen:]
-        logging.debug( "PAM seq: %s of %s" % (genomePamSeq, tSeq))
-        if genomePamSeq in altPamSeqs:
-            minScore = ALTPAMMINSCORE
-        elif genomePamSeq in allPamSeqs:
-            minScore = MINSCORE
-        else:
-            logging.debug("Skipping off-target for %s: %s:%d-%d" % (guideId, chrom, start, end))
-            continue
-
-        logging.debug("off-target minScore = %f" % minScore )
-
-        # check if this match passes the off-target score limit
-        if pamPat=="TTTN":
-            otScore = 0.0
-        else:
-            tSeqNoPam = tSeq[:-pamLen]
-
-            otScore = calcHitScore(guideSeq, tSeqNoPam)
-
-            if otScore < minScore:
-                logging.debug("off-target not accepted")
-                continue
-
-        editDist = getEditDist(guideSeq, tSeqNoPam)
-        guideHitCount = 0
-        guideId = guideId.split(".")[0] # full guide ID looks like s33+.0:A>T
-        name = guideId+"|"+strand+"|"+str(editDist)+"|"+tSeq+"|"+str(guideHitCount)+"|"+str(otScore)
-        row = [chrom, str(start), str(end), name]
-        # this way of collecting the features will remove the duplicates
-        otKey = (chrom, start, end, strand, guideId)
-        logging.debug("off-target key is %s" % str(otKey))
-        offTargets[ otKey ] = row
-
-    for rowKey, row in offTargets.iteritems():
-        tempFh.write("\t".join(row))
-        tempFh.write("\n")
-
-    tempFh.flush()
-
-    # create a tempfile which is moved over upon success
-    # makes sure we do not leave behind a half-written file if 
-    # we crash later
-    tmpFd, tmpAnnotOffsPath = tempfile.mkstemp(dir=tmpDir, prefix="annotOfftargets")
-    tmpFh = open(tmpAnnotOffsPath, "w")
-
-    # get name of file with genome locus names
-    genomeDir = genomesDir # make local var
-    segFname = "%(genomeDir)s/%(genome)s/%(genome)s.segments.bed" % locals()
-
-    # annotate with genome locus names
-    cmd = "$BIN/overlapSelect %(segFname)s %(tempBedPath)s stdout -mergeOutput -selectFmt=bed -inFmt=bed | cut -f1,2,3,4,8 > %(tmpAnnotOffsPath)s" % locals()
-    runCmd(cmd)
-
-    shutil.move(tmpAnnotOffsPath, bedFname)
-    queue.startStep(batchId, "done", "Job completed")
-
-    if DEBUG:
-        logging.info("debug mode: Not deleting %s" % tmpDir)
-    else:
-        shutil.rmtree(tmpDir)
-
 def processSubmission(faFname, genome, pam, bedFname, batchBase, batchId, queue):
     """ search fasta file against genome, filter for pam matches and write to bedFName 
     optionally write status updates to work queue.
     """
+    doEffScoring = False
     if doEffScoring and not cpf1Mode:
         queue.startStep(batchId, "effScores", "Calculating guide efficiency scores")
         createBatchEffScoreTable(batchId)
@@ -1826,9 +1138,9 @@ def processSubmission(faFname, genome, pam, bedFname, batchBase, batchId, queue)
         return
 
     if useBowtie:
-        findOfftargetsBowtie(queue, batchId, batchBase, faFname, genome, pam, bedFname)
+        get_offtargets.findOfftargetsBowtie(queue, batchId, batchBase, faFname, genome, pam, bedFname)
     else:
-        findOfftargetsBwa(queue, batchId, batchBase, faFname, genome, pam, bedFname)
+        get_offtargets.findOfftargetsBwa(queue, batchId, batchBase, faFname, genome, pam, bedFname)
 
     return bedFname
 
@@ -1863,7 +1175,7 @@ allGenomes = None
 
 def readGenomes():
     " return list of all genomes supported "
-    global allGenomes
+    # global allGenomes
     if allGenomes:
         return allGenomes
     genomes = {}
@@ -1904,7 +1216,7 @@ def readBatchParams(batchId):
     jsonFname = batchBase+".json"
     if isfile(jsonFname):
         params = json.load(open(jsonFname))
-        global batchName
+        # global batchName
         batchName = params["batchName"]
         return params["seq"], params["org"], params["pam"], params["posStr"], params["extSeq"]
 
@@ -1949,7 +1261,7 @@ def findAllPams(seq, pam):
     startDict, endSet = findPams(seq, revComp(pam), "-", startDict, endSet)
     return startDict, endSet
 
-def newBatch(batchName, seq, org, pam, skipAlign=False):
+def newBatch(batchName, seq, org, pam, genomesDir,batchDir,skipAlign=False):
     """ obtain a batch ID and write seq/org/pam to their files.
     Return batchId, position string and a 100bp-extended seq, if possible.
     """
@@ -1957,7 +1269,7 @@ def newBatch(batchName, seq, org, pam, skipAlign=False):
     if skipAlign:
         chrom, start, end, strand = None, None, None, None
     else:
-        chrom, start, end, strand = findBestMatch(org, seq)
+        chrom, start, end, strand = findBestMatch(org, seq,genomesDir)
     # define temp file names
     batchBase = join(batchDir, batchId)
     # save input seq, pamSeq, genome, position for primer design later
@@ -1976,7 +1288,7 @@ def newBatch(batchName, seq, org, pam, skipAlign=False):
     # try to get a 100bp-extended version of the input seq
     extSeq = None
     if chrom!=None:
-        extSeq = extendAndGetSeq(org, chrom, start, end, strand)
+        extSeq = extendAndGetSeq(org, genomesDir,chrom, start, end, strand)
         #if extSeq!=None:
             #ofh.write(">extSeq\n%s\n" % (extSeq))
     batchData["extSeq"] = extSeq
@@ -1996,36 +1308,17 @@ def readDbInfo(org):
     return dbInfo
 
 
-def getOfftargets(seq, org, pam, batchId, startDict, queue):
+def getOfftargets(seq, org, pam, batchId,batchDir, startDict, queue):
     """ write guides to fasta and run bwa or use cached results.
     Return name of the BED file with the matches.
     Write progress status updates to queue object.
     """
     batchBase = join(batchDir, batchId)
     otBedFname = batchBase+".bed"
-    flagFile = batchBase+".running"
-
-    if isfile(flagFile):
-       errAbort("This sequence is still being processed. Please wait for ~20 seconds "
-           "and try again, e.g. by reloading this page. If you see this message for "
-           "more than 2-3 minutes, please send an email %s.net. Thanks!" % contactEmail)
-
-    if not isfile(otBedFname) or commandLineMode:
-        # write potential PAM sites to file 
-        faFname = batchBase+".fa"
-        writePamFlank(seq, startDict, pam, faFname)
-        if commandLineMode:
-            processSubmission(faFname, org, pam, otBedFname, batchBase, batchId, queue)
-        else:
-            # umask is not respected by sqlite, bug http://www.mail-archive.com/sqlite-users@sqlite.org/msg59080.html
-            q = JobQueue()
-            ip = os.environ["REMOTE_ADDR"]
-            wasOk = q.addJob("search", batchId, "ip=%s,org=%s,pam=%s" % (ip, org, pam))
-            if not wasOk:
-                #print "CRISPOR job is running..." % batchId
-                pass
-            q.close()
-            return None
+    # write potential PAM sites to file 
+    faFname = batchBase+".fa"
+    writePamFlank(seq, startDict, pam, faFname)
+    processSubmission(faFname, org, pam, otBedFname, batchBase, batchId, queue)
 
     return otBedFname
 
@@ -2163,9 +1456,8 @@ def crisprSearch(params):
             seq = getSeq(params["org"], seq.replace(" ","").replace(",",""))
 
         seq, warnMsg = cleanSeq(seq, org)
-        batchId, position, extSeq = newBatch(newBatchName, seq, org, pam)
-        
-    setupPamInfo(pam)
+        batchId, position, extSeq = newBatch(newBatchName, seq, org, pam) 
+    GUIDELEN,cpf1Mode,addGenePlasmids = common_functions.setupPamInfo(pam)
 
     # check if minFreq was specified
     minFreq = params.get("minFreq", "0.0")
@@ -2394,7 +1686,7 @@ def coordsToPosStr(chrom, start, end, strand):
     locStr = "%s:%d-%d:%s" % (chrom, start, end, strand)
     return locStr
 
-def findBestMatch(genome, seq):
+def findBestMatch(genome, seq,genomesDir):
     """ find best match for input sequence from batchId in genome and return as
     a string "chrom:start-end:strand or None if not found "
     """
@@ -2414,7 +1706,7 @@ def findBestMatch(genome, seq):
 
     genomeDir = genomesDir # make local var
     cmd = "$BIN/bwa bwasw -T 20 %(genomeDir)s/%(genome)s/%(genome)s.fa %(faFname)s > %(samFname)s" % locals()
-    runCmd(cmd)
+    get_offtargets.runCmd(cmd)
 
     chrom, start, end = None, None, None
     for l in open(samFname):
@@ -2487,7 +1779,7 @@ def findGuideSeq(inSeq, pam, pamId):
     the sequence with the pam and its strand.
     """
     startDict, endSet = findAllPams(inSeq, pam)
-    pamInfo = list(flankSeqIter(inSeq, startDict, len(pam), False))
+    pamInfo = list(flankSeqIter(inSeq, startDict, pam, False))
     for guidePamId, pamStart, guideStart, guideStrand, guideSeq, pamSeq in pamInfo:
         if guidePamId!=pamId:
             continue
@@ -2535,7 +1827,7 @@ def findOntargetPos(batchBase, pamId, position):
 
         matchList = [filtMatch]
 
-    global batchName
+    # global batchName
     batchName = batchName.replace(" ", "_")
 
     chrom, start, end, seq, strand, segType, segName, x1Count = \
@@ -2644,7 +1936,7 @@ tmpDirsDelExit = []
 
 def delTmpDirs():
     " signal handler at program exit, to remove registered tmp dirs "
-    global tmpDirsDelExit
+    # global tmpDirsDelExit
     logging.debug("Removing tmpDirs: %s" % ",".join(tmpDirsDelExit))
     for tmpDir in tmpDirsDelExit:
         if isdir(tmpDir):
@@ -2655,6 +1947,7 @@ def delTmpDirs():
 
 def handleOptions(options):
     " set glpbal vars based on options "
+    from constants import DEBUG,doEffScoring,MAXOCC,ALTPAMMINSCORE,maxMMs,useBowtie
     if options.test:
         runTests()
         import doctest
@@ -2662,42 +1955,41 @@ def handleOptions(options):
         sys.exit(0)
 
     if options.debug:
-        global DEBUG
         DEBUG = True
 
     if options.noEffScores:
-        global doEffScoring
         doEffScoring = False
 
     # handle the alignment/filtering options
     if options.maxOcc != None:
-        global MAXOCC
         MAXOCC = options.maxOcc
-        #HIGH_MAXOCC = options.maxOcc
 
     if options.minAltPamScore!=None:
-        global ALTPAMMINSCORE
         ALTPAMMINSCORE = options.minAltPamScore
 
     if options.mismatches:
-        global maxMMs
         maxMMs = options.mismatches
 
     if options.bowtie:
-        global useBowtie
         useBowtie = True
 
     if options.pam:
-        setupPamInfo(options.pam)
+        GUIDELEN,cpf1Mode,addGenePlasmids = common_functions.setupPamInfo(options.pam)
+    argument_list = [DEBUG,doEffScoring,MAXOCC,ALTPAMMINSCORE,maxMMs,
+                            useBowtie,GUIDELEN,cpf1Mode,addGenePlasmids]
+    
+    return argument_list
 
+    
 def mainCommandLine():
     " main entry if called from command line "
-    global commandLineMode
+    # global commandLineMode
     commandLineMode = True
 
-    args, options = common_functions.parseArgs()
-
-    handleOptions(options)
+    args, options = parser_functions.parseArgs()
+    argument_list = handleOptions(options)
+    [DEBUG,doEffScoring,MAXOCC,ALTPAMMINSCORE,maxMMs,
+    useBowtie,GUIDELEN,cpf1Mode,addGenePlasmids] = argument_list
     org, inSeqFname, outGuideFname = args
 
     skipAlign = False
@@ -2706,7 +1998,6 @@ def mainCommandLine():
 
     # different genomes directory?
     if options.genomeDir != None:
-        global genomesDir
         genomesDir = options.genomeDir
 
     # get sequence
@@ -2714,7 +2005,7 @@ def mainCommandLine():
 
     # make a directory for the temp files
     # and put it into a global variable, so all functions will use it
-    global batchDir
+    # global batchDir
     batchDir = tempfile.mkdtemp(dir=TEMPDIR, prefix="crispor")
     logging.debug("Created directory %s" % batchDir)
     if options.debug:
@@ -2738,14 +2029,14 @@ def mainCommandLine():
         # get the other parameters and write to a new batch
         seq = seq.upper()
         pam = options.pam
-        batchId, position, extSeq = newBatch(seqId, seq, org, pam, skipAlign)
+        batchId, position, extSeq = newBatch(seqId, seq, org, pam,genomesDir,batchDir,skipAlign)
         logging.debug("Temporary output directory: %s/%s" % (batchDir, batchId))
 
         if position=="?":
             logging.error("no match found for sequence %s in genome %s" % (inSeqFname, org))
 
         startDict, endSet = findAllPams(seq, pam)
-        otBedFname = getOfftargets(seq, org, pam, batchId, startDict, ConsQueue())
+        otBedFname = getOfftargets(seq, org, pam, batchId, batchDir,startDict, ConsQueue())
         otMatches = parseOfftargets(otBedFname)
 
         if options.noEffScores or cpf1Mode:
