+++ get_offtargets.py
@@ -1,0 +1,218 @@
+import logging,subprocess,shutil
+
+from basic_imports import *
+from constants import *
+import constants as cnst
+import common_functions
+baseDir = cnst.baseDir
+# directory for genomes
+genomesDir = join(baseDir, "genomes")
+
+
+def debug(msg):
+    logging.debug(msg)
+    if DEBUG:
+        print msg
+
+def runCmd(cmd, ignoreExitCode=False):
+    " run shell command, check ret code, replaces BIN and SCRIPTS special variables "
+    cmd = cmd.replace("$BIN", binDir)
+    cmd = cmd.replace("$SCRIPT", scriptDir)
+    cmd = "set -o pipefail; " + cmd
+    debug("Running %s" % cmd)
+    ret = subprocess.call(cmd, shell=True, executable="/bin/bash")
+    if ret!=0 and not ignoreExitCode:
+        if commandLineMode:
+            logging.error("Error: could not run command %s." % cmd)
+            sys.exit(1)
+        else:
+            print "Server error: could not run command %s, error %d.<p>" % (cmd, ret)
+            print "please send us an email, we will fix this error as quickly as possible. %s " % contactEmail
+            sys.exit(0)
+
+
+def findOfftargetsBwa(queue, batchId, batchBase, faFname, genome, pam, bedFname):
+    " align faFname to genome and create matchedBedFname "
+    GUIDELEN,cpf1Mode,addGenePlasmids = common_functions.setupPamInfo(pam)
+    matchesBedFname = batchBase+".matches.bed"
+    saFname = batchBase+".sa"
+    pamLen = len(pam)
+    genomeDir = genomesDir # make var local, see below
+
+    open(matchesBedFname, "w") # truncate to 0 size
+
+    # increase MAXOCC if there is only a single query, but only in CGI mode
+    #if len(parseFasta(open(faFname)))==1 and not commandLineMode:
+        #global MAXOCC
+        #global maxMMs
+        #MAXOCC=max(HIGH_MAXOCC, MAXOCC)
+        #maxMMs=HIGH_maxMMs
+
+    maxDiff = maxMMs
+    queue.startStep(batchId, "bwa", "Alignment of potential guides, mismatches <= %d" % maxDiff)
+    convertMsg = "Converting alignments"
+    seqLen = GUIDELEN
+
+    bwaM = MFAC*MAXOCC # -m is queue size in bwa
+    print ("maxDiff",maxDiff)
+    cmd = "$BIN/bwa aln -o 0 -m %(bwaM)s -n %(maxDiff)d -k %(maxDiff)d -N -l %(seqLen)d %(genomeDir)s/%(genome)s/%(genome)s.fa %(faFname)s > %(saFname)s" % locals()
+    runCmd(cmd)
+
+    queue.startStep(batchId, "saiToBed", convertMsg)
+    maxOcc = MAXOCC # create local var from global
+    # EXTRACTION OF POSITIONS + CONVERSION + SORT/CLIP
+    # the sorting should improve the twoBitToFa runtime
+
+    cmd = "$BIN/bwa samse -n %(maxOcc)d %(genomeDir)s/%(genome)s/%(genome)s.fa %(saFname)s %(faFname)s | $SCRIPT/xa2multi.pl | $SCRIPT/samToBed %(pam)s | sort -k1,1 -k2,2n | $BIN/bedClip stdin %(genomeDir)s/%(genome)s/%(genome)s.sizes stdout >> %(matchesBedFname)s " % locals()
+    runCmd(cmd)
+
+    # arguments: guideSeq, mainPat, altPats, altScore, passX1Score
+    filtMatchesBedFname = batchBase+".filtMatches.bed"
+    queue.startStep(batchId, "filter", "Removing matches without a PAM motif")
+    altPats = ",".join(offtargetPams.get(pam, ["na"]))
+    bedFnameTmp = bedFname+".tmp"
+    altPamMinScore = str(ALTPAMMINSCORE)
+    # EXTRACTION OF SEQUENCES + ANNOTATION
+    # twoBitToFa was 15x slower than python's twobitreader, after markd's fix it should be OK
+    cmd = "$BIN/twoBitToFa %(genomeDir)s/%(genome)s/%(genome)s.2bit stdout -bed=%(matchesBedFname)s | $SCRIPT/filterFaToBed %(faFname)s %(pam)s %(altPats)s %(altPamMinScore)s %(maxOcc)d > %(filtMatchesBedFname)s" % locals()
+    #cmd = "$SCRIPT/twoBitToFaPython %(genomeDir)s/%(genome)s/%(genome)s.2bit %(matchesBedFname)s | $SCRIPT/filterFaToBed %(faFname)s %(pam)s %(altPats)s %(altPamMinScore)s %(maxOcc)d > %(filtMatchesBedFname)s" % locals()
+    runCmd(cmd)
+
+    segFname = "%(genomeDir)s/%(genome)s/%(genome)s.segments.bed" % locals()
+
+    # if we have gene model segments, annotate them, otherwise just use the chrom position
+    if isfile(segFname):
+        queue.startStep(batchId, "genes", "Annotating matches with genes")
+        cmd = "cat %(filtMatchesBedFname)s | $BIN/overlapSelect %(segFname)s stdin stdout -mergeOutput -selectFmt=bed -inFmt=bed | cut -f1,2,3,4,8 2> %(batchBase)s.log > %(bedFnameTmp)s " % locals()
+        runCmd(cmd)
+    else:
+        queue.startStep(batchId, "chromPos", "Annotating matches with chromosome position")
+        annotateBedWithPos(filtMatchesBedFname, bedFnameTmp)
+
+    # make sure the final bed file is never in a half-written state, 
+    # as it is our signal that the job is complete
+    shutil.move(bedFnameTmp, bedFname)
+    queue.startStep(batchId, "done", "Job completed")
+    return bedFname
+
+
+
+def findOfftargetsBowtie(queue, batchId, batchBase, faFname, genome, pamPat, bedFname):
+    " align guides with pam in faFname to genome and write off-targets to bedFname "
+    tmpDir = batchBase+".bowtie.tmp"
+    os.mkdir(tmpDir)
+
+    # make sure this directory gets removed, no matter what
+    global tmpDirsDelExit
+    tmpDirsDelExit.append(tmpDir)
+    if not DEBUG:
+        atexit.register(delTmpDirs)
+
+    # write out the sequences for bowtie
+    queue.startStep(batchId, "seqPrep", "preparing sequences")
+    bwFaFname = abspath(join(tmpDir, "bowtieIn.fa"))
+    guideSeqs, qSeqs, allPamSeqs = writeBowtieSequences(faFname, bwFaFname, pamPat)
+
+    genomePath =  abspath(join(genomesDir, genome, genome))
+    oldCwd = os.getcwd()
+
+    # run bowtie
+    queue.startStep(batchId, "bowtie", "aligning with bowtie")
+    os.chdir(tmpDir) # bowtie writes to hardcoded output filenames with --refout
+    # -v 3 = up to three mismatches
+    # -y   = try hard
+    # -t   = print time it took
+    # -k   = output up to X alignments
+    # -m   = do not output any hit if a read has more than X hits
+    # --max = write all reads that exceed -m to this file
+    # --refout = output in bowtie format, not SAM
+    # --maxbts=2000 maximum number of backtracks
+    # -p 4 = use four threads
+    # --mm = use mmap
+    maxOcc = MAXOCC # meaning in BWA: includes any PAM, in bowtie we have the PAM in the input sequence
+    cmd = "$BIN/bowtie %(genomePath)s -f %(bwFaFname)s  -v 3 -y -t -k %(maxOcc)d -m %(maxOcc)d dummy --max tooManyHits.txt --mm --refout --maxbts=2000 -p 4" % locals()
+    runCmd(cmd)
+    os.chdir(oldCwd)
+
+    queue.startStep(batchId, "parse", "parsing alignments")
+    pamLen = len(pamPat)
+    hits = parseRefout(tmpDir, guideSeqs, pamLen)
+
+    queue.startStep(batchId, "scoreOts", "scoring off-targets")
+    # make the list of alternative PAM sequences
+    altPats = offtargetPams.get(pamPat, [])
+    altPamSeqs = []
+    for altPat in altPats:
+        altPamSeqs.extend(expandIupac(altPat))
+
+    # iterate over bowtie hits and write to a BED file with scores
+    # if the hit looks OK (right PAM + score is high enough)
+    tempBedPath = join(tmpDir, "bowtieHits.bed")
+    tempFh = open(tempBedPath, "w")
+
+    offTargets = {}
+    for guideIdWithMod, chrom, start, end, strand, _, tSeq in hits:
+        guideId = guideIdWithMod.split(".")[0]
+        guideSeq = guideSeqs[guideId]
+        genomePamSeq = tSeq[-pamLen:]
+        logging.debug( "PAM seq: %s of %s" % (genomePamSeq, tSeq))
+        if genomePamSeq in altPamSeqs:
+            minScore = ALTPAMMINSCORE
+        elif genomePamSeq in allPamSeqs:
+            minScore = MINSCORE
+        else:
+            logging.debug("Skipping off-target for %s: %s:%d-%d" % (guideId, chrom, start, end))
+            continue
+
+        logging.debug("off-target minScore = %f" % minScore )
+
+        # check if this match passes the off-target score limit
+        if pamPat=="TTTN":
+            otScore = 0.0
+        else:
+            tSeqNoPam = tSeq[:-pamLen]
+
+            otScore = calcHitScore(guideSeq, tSeqNoPam)
+
+            if otScore < minScore:
+                logging.debug("off-target not accepted")
+                continue
+
+        editDist = getEditDist(guideSeq, tSeqNoPam)
+        guideHitCount = 0
+        guideId = guideId.split(".")[0] # full guide ID looks like s33+.0:A>T
+        name = guideId+"|"+strand+"|"+str(editDist)+"|"+tSeq+"|"+str(guideHitCount)+"|"+str(otScore)
+        row = [chrom, str(start), str(end), name]
+        # this way of collecting the features will remove the duplicates
+        otKey = (chrom, start, end, strand, guideId)
+        logging.debug("off-target key is %s" % str(otKey))
+        offTargets[ otKey ] = row
+
+    for rowKey, row in offTargets.iteritems():
+        tempFh.write("\t".join(row))
+        tempFh.write("\n")
+
+    tempFh.flush()
+
+    # create a tempfile which is moved over upon success
+    # makes sure we do not leave behind a half-written file if 
+    # we crash later
+    tmpFd, tmpAnnotOffsPath = tempfile.mkstemp(dir=tmpDir, prefix="annotOfftargets")
+    tmpFh = open(tmpAnnotOffsPath, "w")
+
+    # get name of file with genome locus names
+    genomeDir = genomesDir # make local var
+    segFname = "%(genomeDir)s/%(genome)s/%(genome)s.segments.bed" % locals()
+
+    # annotate with genome locus names
+    cmd = "$BIN/overlapSelect %(segFname)s %(tempBedPath)s stdout -mergeOutput -selectFmt=bed -inFmt=bed | cut -f1,2,3,4,8 > %(tmpAnnotOffsPath)s" % locals()
+    runCmd(cmd)
+
+    shutil.move(tmpAnnotOffsPath, bedFname)
+    queue.startStep(batchId, "done", "Job completed")
+
+    if DEBUG:
+        logging.info("debug mode: Not deleting %s" % tmpDir)
+    else:
+        shutil.rmtree(tmpDir)
+
